+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * **Mod Support:**
   - Add support for optional menu and intermission sounds.
+  - Add support for ambient sound definitions by SNDINFO lumps.
 
 * **Quality of Life:**
   - Player view no longer abruptly jolts when running across very shallow floor height changes